Fixing the exercise.
