--- flutter_widgets_basic/lib/text_control_button.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class TextControlButton extends StatelessWidget {
-  const TextControlButton({Key? key, required this.callback}) : super(key: key);
-
-  final VoidCallback callback;
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: callback,
-      child: const Text('Change text'),
-    );
-  }
-}
